requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via
    #   aiohttp-sse-client
    #   stellar-sdk
aiohttp-sse-client==0.2.1
    # via stellar-sdk
aiosignal==1.2.0
    # via aiohttp
asgiref==3.4.1
    # via django
async-timeout==4.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   aiohttp-sse-client
black==21.12b0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via pynacl
charset-normalizer==2.0.9
    # via
    #   aiohttp
    #   requests
click==8.0.3
    # via black
django==4.0
    # via
    #   -r requirements.in
    #   django-sslserver
    #   djangorestframework
django-environ==0.8.1
    # via -r requirements.in
django-sslserver==0.22
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
frozenlist==1.2.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.3
    # via
    #   requests
    #   yarl
mnemonic==0.20
    # via stellar-sdk
multidict==5.2.0
    # via
    #   aiohttp
    #   aiohttp-sse-client
    #   yarl
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via black
pycparser==2.21
    # via cffi
pynacl==1.4.0
    # via stellar-sdk
pytz==2021.3
    # via djangorestframework
requests==2.26.0
    # via
    #   stellar-base-sseclient
    #   stellar-sdk
six==1.16.0
    # via
    #   pynacl
    #   stellar-base-sseclient
sqlparse==0.4.2
    # via django
stellar-base-sseclient==0.0.21
    # via stellar-sdk
stellar-sdk==6.1.1
    # via -r requirements.in
toml==0.10.2
    # via stellar-sdk
tomli==1.2.3
    # via black
typeguard==2.13.3
    # via stellar-sdk
typing-extensions==4.0.1
    # via
    #   async-timeout
    #   black
urllib3==1.26.7
    # via
    #   requests
    #   stellar-sdk
yarl==1.7.2
    # via
    #   aiohttp
    #   aiohttp-sse-client
